+++ src/main/java/com/supermartijn642/packedup/BackpackIconRenderer.java
@@ -10,9 +10,9 @@ import net.minecraft.client.renderer.special.SpecialModelRenderer;
 import net.minecraft.world.item.ItemDisplayContext;
 import net.minecraft.world.item.ItemStack;
 import org.jetbrains.annotations.Nullable;
-import org.joml.Vector3f;
+import org.joml.Vector3fc;
 
-import java.util.Set;
+import java.util.function.Consumer;
 
 /**
  * Created 15/01/2025 by SuperMartijn642
@@ -40,7 +40,7 @@ public class BackpackIconRenderer implements ItemModel.Unbaked {
         }
 
         @Override
-        public void getExtents(Set<Vector3f> set){
+        public void getExtents(Consumer<Vector3fc> set){
         }
 
         @Override