+++ changelog.md
@@ -1,2 +1,2 @@
 ### Packed Up 1.1.0
-- Initial release of Packed Up for Minecraft 1.21.9 & 1.21.10
+- Initial release of Packed Up for Minecraft 1.21.11